feat(132645): Agrupando os styles em um arquivo e ajustes nos imports
