--- src/pages/Processos/NovaConvocacaoCandidatos/components/AdicionarEscolaTable/style.ts
@@ -1,7 +1,0 @@
-import { Table, Typography } from "antd";
-import styled from "styled-components";
-
-export const CustomTitle = styled(Typography.Title).attrs({ level: 4 })`
-  margin: 2.375rem 0 1rem 0;
-`;
-
