+++ src/pages/Processos/NovaConvocacaoCandidatos/components/AdicionarNovaEscolaModal/AdicionarNovaEscolaModal.tsx
@@ -6,7 +6,7 @@ const { Text } = Typography;
 
 import { Col, Divider, Input, Row } from "antd";
 import { Controller, useForm } from "react-hook-form";
-import { CustomFormItem } from "./styles";
+import { CustomFormItem } from "../../styles";
 import { Content } from "antd/es/layout/layout";
 import type { IConvocacaoFiltros } from "../../../../../services/resources/convocacao/IConvocacao";
